+++ Sources/Schemas/CallerCredentialTypeEnum.swift
@@ -4,4 +4,5 @@ public enum CallerCredentialTypeEnum: String, Codable, Hashable, CaseIterable, S
     case sessionJwt = "session_jwt"
     case m2MJwt = "m2m_jwt"
     case apiKey = "api_key"
+    case internalDelegationJwt = "internal_delegation_jwt"
 }